refactor: move mixins.py from shared/ to src root
Move infrastructure utilities (TimestampMixin, timezone helpers) from
src/shared/ to src/ to align with existing infrastructure pattern.

src/shared/ now contains only cross-domain business utilities:
- errors.py: Business error codes
- redis_keys.py: Business cache key patterns

Closes #108
