+++ src/shared/__init__.py
@@ -1,5 +1,4 @@
 from .errors import ErrorCode, error_code_to_http_status
-from .mixins import TimestampMixin
 from .redis_keys import redis_keys
 
-__all__ = ["ErrorCode", "error_code_to_http_status", "TimestampMixin", "redis_keys"]
+__all__ = ["ErrorCode", "error_code_to_http_status", "redis_keys"]
