+++ tests/shared/test_mixins.py
@@ -5,7 +5,7 @@ from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
 from sqlalchemy.orm import sessionmaker
 from sqlmodel import Field, SQLModel
 
-from src.shared.mixins import TimestampMixin
+from src.mixins import TimestampMixin
 
 
 class SampleModel(SQLModel, TimestampMixin, table=True):
